lambda-deploy-20251206_143805/requirements.txt
# Lambda-optimized requirements - binary wheels only
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
openai>=1.50.0
fastapi>=0.110.0
uvicorn>=0.30.0
python-multipart>=0.0.20
mangum>=0.19.0
requests>=2.31.0
pandas>=2.0.0
# Playwright with headless mode only
playwright>=1.47.0